% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.swdpower.R
\name{print.swdpower}
\alias{print.swdpower}
\title{Print the results of swdpower}
\usage{
\method{print}{swdpower}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "swdpower"
}
